+++ packages/connect/src/index.ts
@@ -1,11 +1,15 @@
 export {HttpClient} from './http_client';
 export {
     Client,
+    ECSignature,
     FeesRequest,
     FeesResponse,
+    Order,
     OrderbookRequest,
     OrderbookResponse,
     OrdersRequest,
+    SignedOrder,
     TokenPairsItem,
     TokenPairsRequest,
+    TokenTradeInfo,
 } from './types';