+++ packages/connect/src/utils/orderbook_channel_message_parsers.ts
@@ -1,4 +1,3 @@
-import {SignedOrder} from '0x.js';
 import {assert} from '@0xproject/assert';
 import {schemas} from '@0xproject/json-schemas';
 import * as _ from 'lodash';
@@ -6,6 +5,7 @@ import * as _ from 'lodash';
 import {
     OrderbookChannelMessage,
     OrderbookChannelMessageTypes,
+    SignedOrder,
 } from '../types';
 
 import {typeConverters} from './type_converters';