+++ packages/connect/CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+v0.2.0 - _November 29, 2017_
+------------------------
+    * Add SignedOrder and TokenTradeInfo to the public interface
+    * Add ECSignature and Order to the public interface
+    * Remove dependency on 0x.js
+
 v0.1.0 - _November 22, 2017_
 ------------------------
     * Provide a HttpClient class for interacting with standard relayer api compliant HTTP urls